feat: Implement dry-run functionality across processors for safer execution

+++ scripts/processors/10-prefixsuffix.sh
@@ -19,20 +19,31 @@ else
 fi
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-tmp2=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-tmp3=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-tmp4=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-cat $POTFILE | awk -F: '{print $NF}' | tee $tmp &>/dev/null
+tmp=$(dryrun_tempfile prefixsuffix)
+tmp2=$(dryrun_tempfile prefixsuffix)
+tmp3=$(dryrun_tempfile prefixsuffix)
+tmp4=$(dryrun_tempfile prefixsuffix)
 
 # Logic
-if [ "$MACHINE" == "Mac" ]; then
-    cat $tmp | awk -F: '{print $NF}' | sort | tee $tmp2 &>/dev/null && ./scripts/extensions/common-substr-mac -n -p -f $tmp2 >$tmp3 && ./scripts/extensions/common-substr-mac -n -s -f $tmp2 >$tmp4 && rm $tmp2 $tmp
+if dry_run_enabled; then
+    dryrun_note "would extract plaintexts from $POTFILE to $tmp"
+    if [ "$MACHINE" == "Mac" ]; then
+        dryrun_note "would run common-substr-mac prefix/suffix generation into $tmp3 and $tmp4"
+    else
+        dryrun_note "would run common-substr-linux prefix/suffix generation into $tmp3 and $tmp4"
+    fi
 else
-    cat $tmp | awk -F: '{print $NF}' | sort | tee $tmp2 &>/dev/null && ./scripts/extensions/common-substr-linux -n -p -f $tmp2 >$tmp3 && ./scripts/extensions/common-substr-linux -n -s -f $tmp2 >$tmp4 && rm $tmp2 $tmp
+    cat $POTFILE | awk -F: '{print $NF}' | tee $tmp &>/dev/null
+    if [ "$MACHINE" == "Mac" ]; then
+        cat $tmp | awk -F: '{print $NF}' | sort | tee $tmp2 &>/dev/null && ./scripts/extensions/common-substr-mac -n -p -f $tmp2 >$tmp3 && ./scripts/extensions/common-substr-mac -n -s -f $tmp2 >$tmp4 && rm $tmp2 $tmp
+    else
+        cat $tmp | awk -F: '{print $NF}' | sort | tee $tmp2 &>/dev/null && ./scripts/extensions/common-substr-linux -n -p -f $tmp2 >$tmp3 && ./scripts/extensions/common-substr-linux -n -s -f $tmp2 >$tmp4 && rm $tmp2 $tmp
+    fi
 fi
 
 hashcat_base -a1 $tmp3 $tmp4
 hashcat_base -a1 $tmp4 $tmp3
-rm $tmp3 $tmp4
+if ! dry_run_enabled; then
+    rm $tmp3 $tmp4
+fi
 echo -e "\nPrefix suffix processing done\n"
