+++ scripts/processors/11-commonsubstring.sh
@@ -19,18 +19,29 @@ else
 fi
 
 # Temporary Files
-tmp2=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-tmp3=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-tmp4=$(mktemp /tmp/hash-cracker-tmp.XXXX)
-awk -F: '{print $NF}' "$POTFILE" >"$tmp2"
+tmp2=$(dryrun_tempfile commonsubstr)
+tmp3=$(dryrun_tempfile commonsubstr)
+tmp4=$(dryrun_tempfile commonsubstr)
 
 # Logic
-if [ "$MACHINE" == "Mac" ]; then
-    sort "$tmp2" | tee "$tmp3" &>/dev/null && ./scripts/extensions/common-substr-mac -n -f "$tmp3" >"$tmp4" && rm -f -- "$tmp3" "$tmp2"
+if dry_run_enabled; then
+    dryrun_note "would extract plaintexts from $POTFILE to $tmp2"
+    if [ "$MACHINE" == "Mac" ]; then
+        dryrun_note "would run common-substr-mac generation into $tmp4"
+    else
+        dryrun_note "would run common-substr-linux generation into $tmp4"
+    fi
 else
-    sort "$tmp2" | tee "$tmp3" &>/dev/null && ./scripts/extensions/common-substr-linux -n -f "$tmp3" >"$tmp4" && rm -f -- "$tmp3" "$tmp2"
+    awk -F: '{print $NF}' "$POTFILE" >"$tmp2"
+    if [ "$MACHINE" == "Mac" ]; then
+        sort "$tmp2" | tee "$tmp3" &>/dev/null && ./scripts/extensions/common-substr-mac -n -f "$tmp3" >"$tmp4" && rm -f -- "$tmp3" "$tmp2"
+    else
+        sort "$tmp2" | tee "$tmp3" &>/dev/null && ./scripts/extensions/common-substr-linux -n -f "$tmp3" >"$tmp4" && rm -f -- "$tmp3" "$tmp2"
+    fi
 fi
 
 $HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path="$POTFILE" -m"$HASHTYPE" "$HASHLIST" -a1 "$tmp4" "$tmp4"
-rm -f -- "$tmp4"
+if ! dry_run_enabled; then
+    rm -f -- "$tmp4"
+fi
 echo -e "\nSubstring processing done\n"
