+++ scripts/processors/4-word.sh
@@ -23,13 +23,19 @@ source scripts/rules/rules.config
 RULELIST=("$tenKrules" "$NSAKEYv2" "$fordyv1" "$pantag" "$OUTD" "$techtrip2" "$williamsuper" "$digits3" "$dive")
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
+tmp=$(dryrun_tempfile word)
 
 # Logic
 read -p "Enter a word (e.g. company name): " WORD
-echo $WORD >$tmp
+if dry_run_enabled; then
+    dryrun_note "would write custom word input to $tmp"
+else
+    echo $WORD >$tmp
+fi
 for RULE in "${RULELIST[@]}"; do
     hashcat_base $tmp -r $RULE $LOOPBACK
 done
-rm $tmp
+if ! dry_run_enabled; then
+    rm $tmp
+fi
 echo -e "\nWord processing done\n"
