+++ scripts/processors/5-word-bruteforce.sh
@@ -17,16 +17,22 @@ else
 fi
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
+tmp=$(dryrun_tempfile word)
 
 # Logic
 read -p "Enter a word (e.g. company name): " WORD
-echo $WORD >$tmp
+if dry_run_enabled; then
+    dryrun_note "would write custom word input to $tmp"
+else
+    echo $WORD >$tmp
+fi
 hashcat_base -a6 $tmp '?d?d?d?d?d?d?d?d' -i
 hashcat_base -a6 $tmp '?l?l?l?l?l?l' -i
 hashcat_base -a7 '?d?d?d?d?d?d?d?d' $tmp -i
 hashcat_base -a7 '?l?l?l?l?l?l' $tmp -i
 hashcat_base -a6 $tmp '?a?a?a?a?a' -i
 hashcat_base -a7 '?a?a?a?a?a' $tmp -i
-rm $tmp
+if ! dry_run_enabled; then
+    rm $tmp
+fi
 echo -e "\nWord processing done\n"
