+++ scripts/processors/9-iterate.sh
@@ -23,12 +23,18 @@ source scripts/rules/rules.config
 RULELIST=("$rule3" "$robotmyfavorite" "$fbfull" "$tenKrules" "$NSAKEYv2" "$fordyv1" "$pantag" "$OUTD" "$TOXICSP" "$techtrip2" "$williamsuper" "$digits3" "$dive" "$TOXIC10k" "$big" "$generated3" "$huge")
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
+tmp=$(dryrun_tempfile iterate)
 
 # Logic
 for RULE in "${RULELIST[@]}"; do
-    cat $POTFILE | awk -F: '{print $NF}' | sort -u | tee $tmp &>/dev/null
+    if dry_run_enabled; then
+        dryrun_note "would extract unique plaintexts from $POTFILE to $tmp"
+    else
+        cat $POTFILE | awk -F: '{print $NF}' | sort -u | tee $tmp &>/dev/null
+    fi
     hashcat_base $tmp -r $RULE $LOOPBACK
 done
-rm $tmp
+if ! dry_run_enabled; then
+    rm $tmp
+fi
 echo -e "\nIteration processing done\n"
